Refactor board handler and repository methods to include user ID in board operations: Update CreateBoard, GetAllBoards, SaveData, GetBoardByID, DeleteBoardByID, and UpdateBoardByID functions to validate and utilize user ID from context. Modify repository methods to enforce user ownership in board data retrieval and manipulation. Enhance error handling for invalid user IDs and ensure proper image file cleanup on board deletion.

+++ apps/api/internal/repo/board.go
@@ -16,10 +16,10 @@ type BoardRepo struct {
 
 type BoardRepoInterface interface {
 	CreateBoard(board *models.Board) (uuid.UUID, error)
-	GetAllBoards() ([]models.Board, error)
-	GetBoardById(boardId uuid.UUID) (models.Board, error)
-	UpdateBoard(boardId uuid.UUID, board *models.Board) error
-	DeleteBoardByID(boardId uuid.UUID) error
+	GetAllBoards(userID uuid.UUID) ([]models.Board, error)
+	GetBoardById(userID uuid.UUID, boardId uuid.UUID) (models.Board, error)
+	UpdateBoard(userID uuid.UUID, boardId uuid.UUID, board *models.Board) error
+	DeleteBoardByID(userID uuid.UUID, boardId uuid.UUID) error
 }
 
 func NewBoardRepository(db *gorm.DB) BoardRepoInterface {
@@ -37,25 +37,25 @@ func (r *BoardRepo) CreateBoard(board *models.Board) (uuid.UUID, error) {
 }
 
 // GetBoardById returns a board by its ID
-func (r *BoardRepo) GetBoardById(boardId uuid.UUID) (models.Board, error) {
+func (r *BoardRepo) GetBoardById(userID uuid.UUID, boardId uuid.UUID) (models.Board, error) {
 	var board models.Board
-	err := r.db.Where(&models.Board{UUID: boardId}).First(&board).Error
+	err := r.db.Where(&models.Board{UUID: boardId, UserID: userID}).First(&board).Error
 	return board, err
 }
 
 // UpdateBoard updates a board in the database
-func (r *BoardRepo) UpdateBoard(boardId uuid.UUID, board *models.Board) error {
-	return r.db.Model(&models.Board{UUID: boardId}).Updates(board).Error
+func (r *BoardRepo) UpdateBoard(userID uuid.UUID, boardId uuid.UUID, board *models.Board) error {
+	return r.db.Model(&models.Board{UUID: boardId, UserID: userID}).Updates(board).Error
 }
 
 // DeleteBoardByID deletes a board in the database
-func (r *BoardRepo) DeleteBoardByID(boardId uuid.UUID) error {
-	return r.db.Delete(&models.Board{UUID: boardId}).Error
+func (r *BoardRepo) DeleteBoardByID(userID uuid.UUID, boardId uuid.UUID) error {
+	return r.db.Delete(&models.Board{UUID: boardId, UserID: userID}).Error
 }
 
 // GetAllBoards returns all boards in the database
-func (r *BoardRepo) GetAllBoards() ([]models.Board, error) {
+func (r *BoardRepo) GetAllBoards(userID uuid.UUID) ([]models.Board, error) {
 	var boards []models.Board
-	err := r.db.Find(&boards).Error
+	err := r.db.Where(&models.Board{UserID: userID}).Find(&boards).Error
 	return boards, err
 }
